Original Code/requirements.txt
mrc
absl-py==0.13.0
argon2-cffi==18.3.0
asn1crypto==1.4.0
astunparse==1.6.3
attrs==20.3.0
backcall==0.2.0
bcrypt==3.1.7
beautifulsoup4==4.9.3
bleach==3.2.1
blinker==1.4
cachetools==4.2.2
certifi==2020.6.20
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cryptography==3.3.2
cycler==0.10.0
dbus-python==1.2.16
decorator==4.4.2
defusedxml==0.6.0
distro==1.5.0
distro-info==1.0
Django==2.2.20
docutils==0.16
duplicity==0.8.17
entrypoints==0.3
evdev==1.4.0
fasteners==0.14.1
flatbuffers==1.12
future==0.18.2
gast==0.4.0
grpcio==1.34.1
h5py==3.1.0
html5lib==1.1
idna==2.10
image==1.5.33
imageio==2.9.0
importlib-metadata==1.6.0
ipykernel==5.4.3
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==6.0.0
JACK-Client==0.5.3
jedi==0.18.0
jeepney==0.6.0
Jinja2==2.11.2
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-console==6.2.0
jupyter-core==4.7.1
kazam==1.4.5
Keras==2.4.3
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
keyring==22.2.0
Kivy==1.11.1
kiwisolver==1.3.1
language-selector==0.1
launchpadlib==1.10.13
libtiff==0.4.2
lockfile==0.12.2
louis==3.16.0
lxml==4.6.3
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.2.2
meld==3.20.3
mercurial==5.6.1
mistune==0.8.4
monotonic==1.5
more-itertools==4.2.0
mpmath==0.0.0
nbconvert==5.6.1
nbformat==5.1.2
netifaces==0.10.9
networkx==2.4
notebook==6.2.0
numpy==1.19.5
oauth==1.0.1
oauthlib==3.1.0
olefile==0.46
opencv-python==4.5.1.48
opt-einsum==3.3.0
packaging==20.9
pandas==1.0.5
pandocfilters==1.4.3
paramiko==2.7.2
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
ply==3.11
prometheus-client==0.9.0
prompt-toolkit==3.0.14
protobuf==3.12.4
pyalsaaudio==0.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycparser==2.20
pycrypto==2.6.1
pycryptodomex==3.9.7
pycups==2.0.1
Pygments==2.7.1
PyGObject==3.38.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.4.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.15.5
python-apt===2.2.0-ubuntu0.21.04.1
python-augeas==0.5.0
python-dateutil==2.8.1
python-debian==0.1.39
python-distutils-extra==2.39
python-magic==0.4.20
pytz==2021.1
PyWavelets==1.1.1
pyxdg==0.27
PyYAML==5.3.1
pyzmq==20.0.0
reportlab==3.5.66
requests==2.25.1
requests-oauthlib==1.3.0
roman==2.0.0
rsa==4.7.2
s3cmd==2.1.0
scikit-image==0.17.2
scipy==1.5.4
SecretStorage==3.3.1
Send2Trash==1.6.0b1
setproctitle==1.2.1
simplejson==3.17.2
six==1.15.0
soupsieve==2.2
sqlparse==0.4.1
system-service==0.3
systemd-python==234
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
terminado==0.9.2
testpath==0.4.4
tifffile==2021.7.2
tkintertable==1.3.3
tornado==6.1
traitlets==5.0.5
typing-extensions==3.7.4.3
ubuntu-advantage-tools==27.0
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.2
usb-creator==0.3.7
vboxapi==1.0
wadllib==1.3.5
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==2.0.0
wrapt==1.12.1
xdg==5
YeastAnalysisTool==0.1
zipp==1.0.0
zope.interface==5.2.0
opencv-rolling-ball==1.0.1
